For organisation purposesthe rest controller have been refactored.
