+++ src/main/java/com/itwins/foreverbio/controllers/RestController.java
@@ -26,60 +26,7 @@ public class RestController {
 	}
 
 	
-	/*This method take as an argument a json and it's the body of the
-	 * http request then it creates the user object out of it
-	 * and then puts it into the databse.
-	 * for the request body response , we have a json also with the status
-	 * code and the description.
-	 * 
-	 * The json should be in this format :
-	 * 	{
-			"email": "ahmedbougern998@gmail.com",
-			"firstname": "Amine",
-			"lastname": "Elmouradi",
-			"age": 21,
-			"password": "hello world 123"
-		}
-	 * 
-	 * 
-	 */
 
-	@PostMapping("/addUser")
-	public String addUser(@RequestBody Map<String, Object> userMap) {
-		User user = new User(userMap);
-		userService.saveUser(user);
-		return "{"
-				+ "\"statusCode\": 1,"
-				+ "\"description\": \"User saved successfully.\""
-				+ "}";
-	}
-	
-	/* This method take as an argument a json that contains the mail and the password
-	 * Then checks it on the database, creates an object out of it if it exists.
-	 * Then it sends back a json as response with the status code and the user id.
-	 * 
-	 */
-	
-	@PostMapping("/loginUser")
-	public String loginUser(@RequestBody Map<String, Object> map) {
-
-		User user = userService.findUserByEmailAndPassword((String)map.get("email"), (String) map.get("password"));
-		if (user == null) {
-			return "{"
-					+ "\"statusCode\": 0,"
-					+ "\"description\": \"Wrong credentiels.\""
-					+ "}";
-		}
-		return "{"
-				+ "\"statusCode\": 1,"
-				+ "\"description\": \"User logged in successfully.\","
-				+ "\"id\": \"" + user.getId() +"\","
-				+ "\"email\": \"" + user.getEmail() +"\","
-				+ "\"firstname\": \"" + user.getFirstName() +"\","
-				+ "\"lastname\": \"" + user.getLastName() +"\","
-				+ "\"age\": " + user.getAge() +""
-				+ "}";
-	}
 
 	@GetMapping("/about")
 	public String about() {
