+++ src/main/resources/application.properties
@@ -4,5 +4,5 @@ spring.mvc.view.suffix=.jsp
 
 spring.datasource.url=jdbc:mysql://localhost/foreverbio
 spring.datasource.username=root
-spring.datasource.password=
-spring.datasource.driver-class-name=com.mysql.jdbc.Driver
+spring.datasource.password=helloworld?123
+spring.datasource.driver-class-name=com.mysql.cj.jdbc.Driver
